improve performance by %40 by: using size_t instead of uint64_t increasing direct cache index size allocating direct cache index on dram reducing lru changes

+++ lib/VMM/vmm.h
@@ -22,11 +22,11 @@ extern "C" {
 
 typedef struct VMM_t
 {
-    uint64_t page_size;
-    uint64_t number_of_pages;
+    size_t page_size;
+    size_t number_of_pages;
 
     // list_t *pagetable;  // vmTable_t
-    uint64_t pagetable_size;
+    size_t pagetable_size;
     FILE *backing_store;
 
     char filename[255];
@@ -36,29 +36,29 @@ typedef struct VMM_t
     direct_cache_t * direct_cache;
     page_cache_t * page_cache;
 #ifdef VMM_DEBUG
-    uint64_t reads;
-    uint64_t read_bytes;
-    uint64_t writes;
-    uint64_t write_bytes;
+    size_t reads;
+    size_t read_bytes;
+    size_t writes;
+    size_t write_bytes;
 
-    uint64_t store_writes;
-    uint64_t store_write_bytes;
-    uint64_t store_write_time;
-    uint64_t store_reads;
-    uint64_t store_read_bytes;
-    uint64_t store_read_time;
+    size_t store_writes;
+    size_t store_write_bytes;
+    size_t store_write_time;
+    size_t store_reads;
+    size_t store_read_bytes;
+    size_t store_read_time;
 
-    uint64_t page_faults;
+    size_t page_faults;
 #endif
 } VMM_t;
 
-VMM_t *vmm_create(const char *pagefile, uint64_t maximum_size, uint64_t page_size, uint64_t number_of_pages, size_t maximum_himem_blocks);
+VMM_t *vmm_create(const char *pagefile, size_t maximum_size, size_t page_size, size_t number_of_pages, size_t maximum_himem_blocks);
 
 void vmm_destroy(VMM_t *vmm);
 
-void vmm_read(VMM_t *vmm, uint64_t addr, void *target, uint64_t len);
+void vmm_read(VMM_t *vmm, size_t addr, void *target, size_t len);
 
-void vmm_write(VMM_t *vmm, const uint64_t addr, const void *source, const uint64_t len);
+void vmm_write(VMM_t *vmm, const size_t addr, const void *source, const size_t len);
 
 void vmm_flush(VMM_t *vmm);
